backend/requirements.txt
fastapi>=0.118.3
uvicorn==0.23.2
python-dotenv==1.1.1
psutil==5.9.5
ollama==0.1.9
sentence-transformers==2.2.2
faiss-cpu==1.12.0
langchain-community==0.3.29
langchain==0.3.27
chromadb==0.5.17
openai-whisper
moviepy==1.0.3
pillow==11.3.0
pytesseract==0.3.13
python-docx==1.2.0
pandas==2.3.2
PyPDF2==3.0.1
requests==2.32.5
sseclient-py==1.8.0
huggingface-hub==0.33.5
python-multipart
python-docx
pandas
sse-starlette
pydantic==2.11.3
